move to canRenderApp.ts
